@@ -19,54 +19,49 @@
     "test": "ng test --watch false --browsers=ChromeHeadlessCustom",
     "test:watch": "ng test",
     "lint": "ng lint",
-    "e2e": "ng ../e2e"
+    "e2e": "ng ../e2e",
+    "prepare": "husky install"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.5",
-    "@angular/cli": "~13.2.5",
-    "@angular/compiler-cli": "~13.2.4",
-    "@angular/language-service": "~13.2.4",
+    "@angular-devkit/build-angular": "~13.3.3",
+    "@angular/cli": "~13.3.3",
+    "@angular/compiler-cli": "~13.3.4",
+    "@angular/language-service": "~13.3.4",
     "@types/file-saver": "~2.0.2",
-    "@types/jasmine": "~3.10.3",
+    "@types/jasmine": "^4.0.3",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~4.0.1",
+    "husky": "^7.0.0",
+    "jasmine-core": "^4.1.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "^6.3.4",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
+    "karma-jasmine": "^5.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "prettier": "2.5.1",
-    "pretty-quick": "^3.1.0",
+    "prettier": "^2.6.2",
+    "pretty-quick": "^3.1.3",
     "protractor": "~7.0.0",
-    "ts-node": "~10.5.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.5.5"
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.3"
   },
   "dependencies": {
-    "@angular/animations": "~13.2.4",
-    "@angular/common": "~13.2.4",
-    "@angular/compiler": "~13.2.4",
-    "@angular/core": "~13.2.4",
-    "@angular/forms": "~13.2.4",
-    "@angular/platform-browser": "~13.2.4",
-    "@angular/platform-browser-dynamic": "~13.2.4",
-    "@angular/router": "~13.2.4",
+    "@angular/animations": "~13.3.4",
+    "@angular/common": "~13.3.4",
+    "@angular/compiler": "~13.3.4",
+    "@angular/core": "~13.3.4",
+    "@angular/forms": "~13.3.4",
+    "@angular/platform-browser": "~13.3.4",
+    "@angular/platform-browser-dynamic": "~13.3.4",
+    "@angular/router": "~13.3.4",
     "@fortawesome/angular-fontawesome": "^0.10.1",
-    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.0.0",
     "chart.js": "^3.7.1",
     "file-saver": "^2.0.5",
+    "lodash": "^4.17.21",
     "ng2-charts": "^3.0.8",
     "rxjs": "~7.5.4",
-    "lodash": "^4.17.21",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "pretty-quick --staged"
-    }
   }
 }
