@@ -1,31 +1,28 @@
-import {ChangeDetectionStrategy, Component, OnDestroy, OnInit} from '@angular/core';
-import {Observable, Subject} from "rxjs";
-import {Project} from "../../../projects/project";
-import {ProjectCount, SummaryService} from "../../services/summary.service";
-import {ProjectsService} from "../../../projects/services/projects.service";
-import {HeadingService} from "../../../services/heading.service";
-import {takeUntil} from "rxjs/operators";
+import { Component, OnInit } from '@angular/core';
+import { Observable } from 'rxjs';
+import { Project } from '../../../projects/project';
+import { ProjectCount, SummaryService } from '../../services/summary.service';
+import { ProjectsService } from '../../../projects/services/projects.service';
+import { HeadingService } from '../../../services/heading.service';
 
 @Component({
-             selector: 'app-summary',
-             templateUrl: './summary.component.html',
-             styleUrls: ['./summary.component.css'],
-             providers: [SummaryService, ProjectsService],
-
-           })
+  selector: 'app-summary',
+  templateUrl: './summary.component.html',
+  styleUrls: ['./summary.component.css'],
+  providers: [SummaryService, ProjectsService],
+})
 export class SummaryComponent implements OnInit {
   projects$: Observable<Project[]>;
   projectsByOrgan$: Observable<ProjectCount[]>;
   projectsByTech$: Observable<ProjectCount[]>;
   cellCount$: Observable<number>;
 
-  constructor(private projectService: ProjectsService,
-              private summaryService: SummaryService,
-              private headingService: HeadingService) {
-    this.headingService.setTitle(
-      'Summary',
-      'Projects Summary Statistics'
-    );
+  constructor(
+    private projectService: ProjectsService,
+    private summaryService: SummaryService,
+    private headingService: HeadingService
+  ) {
+    this.headingService.setTitle('Summary', 'Projects Summary Statistics');
     this.headingService.setBreadcrumbs('Summary');
   }
 
@@ -35,5 +32,4 @@ export class SummaryComponent implements OnInit {
     this.projectsByTech$ = this.summaryService.projectsByTech$;
     this.cellCount$ = this.summaryService.cellCount$;
   }
-
 }
