@@ -1,7 +1,7 @@
-import {Injectable, OnDestroy} from '@angular/core';
-import {map} from 'rxjs/operators';
-import {Subject} from "rxjs";
-import {ProjectsService} from '../../projects/services/projects.service';
+import { Injectable, OnDestroy } from '@angular/core';
+import { map } from 'rxjs/operators';
+import { Subject } from 'rxjs';
+import { ProjectsService } from '../../projects/services/projects.service';
 
 export interface ProjectCount {
   group: string;
@@ -11,7 +11,6 @@ export interface ProjectCount {
 
 @Injectable()
 export class SummaryService implements OnDestroy {
-
   private projectsByOrgan = new Subject<ProjectCount[]>();
   projectsByOrgan$ = this.projectsByOrgan.asObservable();
 
@@ -24,30 +23,33 @@ export class SummaryService implements OnDestroy {
   constructor(private projectService: ProjectsService) {
     this.groupProjectsByKey('organs', this.projectsByOrgan);
     this.groupProjectsByKey('technologies', this.projectsByTech);
-    this.projectService.getAllProjects()
-        .pipe(map(projects => projects.map(x => x.cellCount)))
-        .subscribe(projects => this.cellCount.next(projects.reduce((acc, val) => acc + val, 0)));
+    this.projectService
+      .getAllProjects()
+      .pipe(map((projects) => projects.map((x) => x.cellCount)))
+      .subscribe((projects) =>
+        this.cellCount.next(projects.reduce((acc, val) => acc + val, 0))
+      );
   }
 
-  groupProjectsByKey(key: string, subject: Subject<any>) : void  {
-    this.projectService.getAllProjects()
-        .pipe(
-          map(projects => {
-            let groupedProjects = this.groupListByKey(projects, key);
-            let data: ProjectCount[] = [];
-            this.sortByValue(groupedProjects)
-                ?.forEach(([key, value]) => {
-                  data.push({
-                              count: value.count,
-                              cellCount: value.cellCount,
-                              group: key
-                            });
-                });
-            return data;
-          })
-        )
-        .subscribe(x => subject.next(x),
-                   x => subject.error(x));
+  groupProjectsByKey(key: string, subject: Subject<any>): void {
+    this.projectService
+      .getAllProjects()
+      .pipe(
+        map((projects) => {
+          let groupedProjects = this.groupListByKey(projects, key);
+          return SummaryService.sortByValue(groupedProjects)?.map<ProjectCount>(
+            ([key, value]) => ({
+              count: value.count,
+              cellCount: value.cellCount,
+              group: key,
+            })
+          );
+        })
+      )
+      .subscribe({
+        next: (x) => subject.next(x),
+        error: (x) => subject.error(x),
+      });
   }
 
   ngOnDestroy() {
@@ -55,20 +57,21 @@ export class SummaryService implements OnDestroy {
     this.projectsByTech.complete();
   }
 
-  private sortByValue(groupedProjects: { [p: string]: any }) {
-    return Object.entries(groupedProjects)
-                 ?.sort(([k1, v1], [k2, v2]) => v2.count - v1.count);
+  private static sortByValue(groupedProjects: { [p: string]: any }) {
+    return Object.entries(groupedProjects)?.sort(
+      ([k1, v1], [, v2]) => v2.count - v1.count
+    );
   }
 
   private groupListByKey(list: any[], groupKey: string) {
     return list?.reduce((acc, val) => {
-      val[groupKey].forEach((key => {
+      val[groupKey].forEach((key) => {
         if (!acc[key]) {
-          acc[key] = {count: 0, cellCount: 0};
+          acc[key] = { count: 0, cellCount: 0 };
         }
         acc[key].count += 1;
         acc[key].cellCount += val.cellCount;
-      }));
+      });
       return acc;
     }, {});
   }
