@@ -6,15 +6,17 @@ export class ProjectsTsvService {
   }
 
   private static projectsAsTsvArray(projects: Project[], columns: object) {
-    const tsvArray = projects.map(
-      (project: Project) => this.projectAsTsvString(project, Object.keys(columns))
+    const tsvArray = projects.map((project: Project) =>
+      this.projectAsTsvString(project, Object.keys(columns))
     );
     tsvArray.unshift(Object.values(columns).join('\t'));
     return tsvArray;
   }
 
   private static projectAsTsvString(project: Project, keys) {
-    return keys.map(key => ProjectsTsvService.flattenProjectField(key, project[key])).join('\t');
+    return keys
+      .map((key) => ProjectsTsvService.flattenProjectField(key, project[key]))
+      .join('\t');
   }
 
   private static flattenProjectField(key: string, value) {
@@ -22,10 +24,14 @@ export class ProjectsTsvService {
       return '';
     }
     if (key === 'authors') {
-      return value.map(author => author.formattedName).join(', ');
+      return value.map((author) => author.formattedName).join(', ');
     }
     if (key === 'publications') {
-      return value.map(publication => `[${publication.journalTitle}](${publication.url})`).join(', ');
+      return value
+        .map(
+          (publication) => `[${publication.journalTitle}](${publication.url})`
+        )
+        .join(', ');
     }
     if (Array.isArray(value)) {
       return value.join(', ');
