@@ -1,19 +1,16 @@
-import {Component, Input, OnInit} from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-mini-card',
   templateUrl: './mini-card.component.html',
-  styleUrls: ['./mini-card.component.css']
+  styleUrls: ['./mini-card.component.css'],
 })
 export class MiniCardComponent implements OnInit {
+  @Input() title: string;
+  @Input() value: string;
+  @Input() icon: string;
 
-  @Input() title:string;
-  @Input() value:string;
-  @Input() icon:string;
+  constructor() {}
 
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
+  ngOnInit(): void {}
 }
