@@ -1,27 +1,21 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
-import {SummaryComponent} from './summary.component';
-import {ProjectsService} from "../../../projects/services/projects.service";
-import {SummaryService} from "../../services/summary.service";
-import {HttpClientTestingModule} from "@angular/common/http/testing";
-import {HeadingService} from "../../../services/heading.service";
+import { SummaryComponent } from './summary.component';
+import { ProjectsService } from '../../../projects/services/projects.service';
+import { SummaryService } from '../../services/summary.service';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { HeadingService } from '../../../services/heading.service';
 
 describe('SummaryComponent', () => {
-
   let component: SummaryComponent;
   let fixture: ComponentFixture<SummaryComponent>;
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      imports:[HttpClientTestingModule],
-                                     providers: [
-                                       HeadingService,
-                                       ProjectsService,
-                                       SummaryService
-                                     ],
-                                     declarations: [SummaryComponent]
-                                   })
-           .compileComponents();
+      imports: [HttpClientTestingModule],
+      providers: [HeadingService, ProjectsService, SummaryService],
+      declarations: [SummaryComponent],
+    }).compileComponents();
   }));
 
   beforeEach(() => {
@@ -31,11 +25,8 @@ describe('SummaryComponent', () => {
   });
 
   it('should create', () => {
-    expect(component)
-      .toBeTruthy();
+    expect(component).toBeTruthy();
   });
 
-  it('should ', function () {
-
-  });
+  it('should ', function () {});
 });
